Refactorized activities repositories/services/controllers. Added routes for sublists view (e.g attendances by activityId). Added routes for specific view (by composite id).
Added Added public composite ids for activities. Updated json serializers. Refactorized Students/Lecturers controllers/services to work with DTOs. Added related Dtos. Added modelMapper. Updated database script.

+++ src/main/java/app/controller/course/CourseController.java
@@ -37,7 +37,7 @@ public class CourseController extends BaseController {
     @RequestMapping(value = "/courses/{id}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
     public
     @ResponseBody
-    ResponseEntity<Course> getCourseById(@PathVariable("id") int id) {
+    ResponseEntity<Course> getCourseById(@PathVariable("id") long id) {
         Course course = courseService.findById(id);
         if (course == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
